garbage/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
geoalchemy2
psycopg2-binary
pydantic
python-dotenv
httpx
